backend/requirements.txt
# Django and DRF
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15

# Database
psycopg2-binary>=2.9.9,<3.0  # 開発環境用
# psycopg2>=2.9.9,<3.0  # 本番環境ではこちらを使用

# CORS 設定
django-cors-headers>=4.3.0,<4.4

# Authentication
djangorestframework-simplejwt>=5.3.0,<6.0  # JWT 認証